requirements.txt
openai>=1.79.0
openai-agents>=0.10.0,<1.0.0
python-dotenv>=1.0.0
